build: use rollup to generate build
use `npm install` to get dependencies.
use `npm start` to view code in a local web server.
use `npm test` to run lint checks.
