+++ README.md
@@ -15,10 +15,10 @@ So far tested in FF, Chrome, Safari and Edge (Windows, Mac, Linux, IOS & Android
  * fr-FR
  * es-ES
  * nl-NL
- 
+
 ## Screenshots
 
-![Card webcomponent](/assets/cardwebcomponent.png "card in en-US")
+![Card webcomponent](public/assets/cardwebcomponent.png "card in en-US")
 
 ## How it works
 
@@ -35,8 +35,7 @@ Define a SimpleCMS portlet with HTML content like the following:
 
 ```HTML
 <my-card id="what-is-uportal-i18n" messagesPath="@uPortal/card/" cssPath="@uPortal/card/css"></my-card>      
-      
+
 <script src='https://cdnjs.cloudflare.com/ajax/libs/webcomponentsjs/1.2.0/webcomponents-lite.js' defer></script>
-<script src='https://unpkg.com/hyperhtml@latest/min.js' defer></script>
 <script  src="@uPortal/card/js/index.js" defer></script>
 ```
