+++ babel.config.js
@@ -1,0 +1,6 @@
+module.exports = function(api) {
+  api.cache.never();
+  return {
+    presets: ['@babel/preset-env'],
+  };
+};
