+++ static/i18n/jsonschema.json
