+++ static/index.html
@@ -9,26 +9,25 @@
       <meta name="apple-mobile-web-app-capable" content="yes"/>
       <meta name="apple-mobile-web-app-status-bar-style" content="black"/>
       <meta name="description" content="POC webcomponents"/>
-      
+
 
 
 
 <!-- Latest compiled and minified CSS -->
       <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.4.2/css/all.css" integrity="sha384-/rXc/GQVaYpyDdyxK+ecHPVYJSN9bmVFBvjA/9eOB+pb3F2w2N6fc5qB9Ew5yIns" crossorigin="anonymous">
       <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous" media="all">
-     
 
 
-     
+
+
 
    </head>
 
    <body>
-      <my-card id="what-is-uportal-i18n" messagesPath="./" cssPath="css"></my-card>      
-      
+      <my-card id="what-is-uportal-i18n" messagesPath="./" cssPath="css"></my-card>
+
       <script src='https://cdnjs.cloudflare.com/ajax/libs/webcomponentsjs/1.2.0/webcomponents-lite.js' defer></script>
-      <script src='https://unpkg.com/hyperhtml@latest/min.js' defer></script>
-      <script  src="js/index.js" defer></script>
+      <script src="my-card.umd.js" defer></script>
 
    </body>
